Updating testParametersContainingSubDefinitions
I removed the `"foo"=>"bar"` part of the definitions array.
A DefinitionProvider should return an array of `DefinitionInterface`, indexed by key. `"bar"` is not a valid `DefinitionInterface`.

Note: the `ArrayDefinitionProvider` used in this test is NOT the one from Assembly (that wraps values not implementing the `DefinitionInterface` into `ParameterDefinition` objects). It is one I wrote for the tests that does pass the array "as is".